Refactored codebase for modularity, readability, and scalability via manager classes
